Various migration fixes
